moved the attribute solver into terraform package
